spark/requirements.txt
psycopg2-binary==2.9.9
kafka-python==2.0.2
py4j==0.10.9.7
pyspark==3.5.0
tenacity==8.2.3
prometheus-client==0.20.0
pytest==8.0.0
pytest-mock==3.12.0
pytest-asyncio==0.23.5
